Use '_' in location range.
